src/main/java/ru/isu/productsaccounting/service/ProductService.java
package ru.isu.productsaccounting.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import ru.isu.productsaccounting.exception.InvalidReference;
import ru.isu.productsaccounting.model.Product;
import ru.isu.productsaccounting.repository.ProductRepository;

import java.util.List;
import java.util.Optional;

@Service
public class ProductService {

    private final ProductRepository productRepository;

    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    public List<Product> getAllProducts() {
        return productRepository.findAll();
    }

    public void saveProduct(Product product) {
        productRepository.save(product);
    }

    public void deleteProduct(Product product) {
        productRepository.delete(product);
    }

    public Optional<Product> findProductByName(String name) {
        return productRepository.findProductByName(name);
    }
}
